+++ include/rt/light.hpp
@@ -54,7 +54,7 @@ namespace rt
 		* Scene is passed so that the camera position, and the objects can be used for computing the specular component
 		* of direct illumination and the shadow computations.
 		**/ 
-		virtual color_t direct(const Vector3f& hitpt, const Vector3f& normal, const material_t* mat, const scene_t* scn) const = 0;
+		virtual color_t direct(const Vector3d& hitpt, const Vector3d& normal, const material_t* mat, const scene_t* scn) const = 0;
 
 		/// Prints information about the light to the stream.
 		virtual void print(std::ostream &stream) const = 0;
@@ -67,15 +67,15 @@ namespace rt
 	{
 	private:
 		/// Position of the light
-		Vector3f pos;
+		Vector3d pos;
 		/// Color of the light. This can be thought of as radiance emitted by the light source.
-		Vector3f col;
+		Vector3d col;
 		/// An ambient coefficient. Modulate col with ka to get ambient component of illumination.
-		float ka;
+		double ka;
 
 	public:
 		/// Constructor
-		point_light_t(const Vector3f& _pos, const Vector3f& _col, const float _ka);
+		point_light_t(const Vector3d& _pos, const Vector3d& _col, const double _ka);
 		/// Destructor
 		virtual ~point_light_t();
 
@@ -84,7 +84,7 @@ namespace rt
 		* Scene is passed so that the camera position, and the objects can be used for computing the specular component
 		* of direct illumination and the shadow computations.
 		**/ 
-		virtual color_t direct(const Vector3f& hitpt, const Vector3f& normal, const material_t* mat, const scene_t* scn) const;
+		virtual color_t direct(const Vector3d& hitpt, const Vector3d& normal, const material_t* mat, const scene_t* scn) const;
 
 		/// Prints information about the light to the stream.
 		virtual void print(std::ostream &stream) const;
